+++ CHANGELOG.md
@@ -46,6 +46,7 @@ This page lists the main changes made to Myokit in each release.
   - [#933](https://github.com/myokit/myokit/pull/933) Made `myokit.step` base its error classification (large/small/none) on numerics, not representation.
   - [#947](https://github.com/myokit/myokit/pull/947) Fixed minor memory leak (reference counting issue) when using fixed form protocols.
   - [#967](https://github.com/myokit/myokit/pull/967) Myokit no longer calls `logging.basicConfig()` when used as a library.
+  - [#977](https://github.com/myokit/myokit/pull/977) `myokit.save()` now raises a `ValueError` if `model`, `protocol`, and `script` are all None.
 
 ## [1.33.9] - 2022-11-08
 - Fixed